tas/fastpath: fix incorrect remote window logic on retx
rx_remote_avail stores the available rx buffer on the remote
peer. This field is determined by the window as reported
by the remote peer in the TCP segment header.

Currently on retransmission, the code reduces the
remote window size by tx_sent (data marked for retransmission),
while also setting tx_sent=0. This leads to understimation
of the flow control window in tcp_txavail().

Fix this in flow_reset_retransmit().

Signed-off-by: Rajath Shashidhara <rajaths@cs.washington.edu>